backend/requirements.txt
Django==2.2.16
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.1.0
PyJWT==2.1.0
